refactor(skills-editor): extract all render functions to renderers.js
