+++ nx2/blocks/skills-editor/nx-skills-editor.js
@@ -31,27 +31,24 @@ import {
   DA_SKILLS_EDITOR_SUGGESTION_HANDOFF,
   DA_SKILLS_EDITOR_CLEAR_FORM_FROM_CHAT,
   DA_SKILLS_EDITOR_FORM_DISMISS,
-  DA_SKILLS_EDITOR_PROMPT_ADD_TO_CHAT,
   DA_SKILLS_EDITOR_PROMPT_SEND,
   DA_SKILLS_LAB_SUGGESTION_HANDOFF,
   DA_SKILLS_LAB_CLEAR_FORM_FROM_CHAT,
   DA_SKILLS_LAB_FORM_DISMISS,
-  DA_SKILLS_LAB_PROMPT_ADD_TO_CHAT,
   DA_SKILLS_LAB_PROMPT_SEND,
 } from './skills-editor-api.js';
 import {
   BUILTIN_AGENTS,
-  BUILTIN_MCP_SERVERS,
-  BUILTIN_TOOL_DETAILS,
   BUILTIN_TOOL_IDS,
-  CATALOG_TABS,
-  CATEGORY_OPTIONS,
   FRESH_FORM_STATE,
-  KNOWN_CATEGORY_CLASSES,
   STATUS,
   STATUS_TYPE,
-  TAB_ACTIONS,
 } from './constants.js';
+import {
+  renderChatDrawer,
+  renderListCol,
+  renderEditorPanel,
+} from './renderers.js';
 
 const styles = await loadStyle(import.meta.url);
 
@@ -509,12 +506,6 @@ class NxSkillsEditor extends LitElement {
     if (type === 'ok') {
       this._statusTimer = setTimeout(() => { this._statusMsg = ''; }, 3000);
     }
-  }
-
-  _msgClass() {
-    if (this._statusType === STATUS_TYPE.ERR) return 'msg-err';
-    if (this._statusType === STATUS_TYPE.WARN) return 'msg-warn';
-    return 'msg-ok';
   }
 
   // ─── dirty form tracking ─────────────────────────────────────────────────
@@ -1298,911 +1289,10 @@ class NxSkillsEditor extends LitElement {
           <span class="refresh-indicator-track"><span class="refresh-indicator-bar"></span></span>
         </div>
       ` : nothing}
-      ${this._renderChatDrawer()}
-      ${this._renderListCol()}
-      ${this._renderEditorPanel()}
+      ${renderChatDrawer(this)}
+      ${renderListCol(this)}
+      ${renderEditorPanel(this)}
     </div>`;
-  }
-
-  // ─── render: list column ──────────────────────────────────────────────────
-
-  _renderListCol() {
-    const tab = this._catalogTab;
-    const showSearch = ['skills', 'prompts', 'mcps'].includes(tab);
-
-    return html`
-      <div class="col col-list" role="region" aria-label="Catalog">
-        <div class="list-header">
-          <nx-tabs
-            .items=${CATALOG_TABS}
-            .active=${tab}
-            @tab-change=${(e) => this._onTabChange(e.detail.id)}
-          ></nx-tabs>
-          <div class="list-actions-row">
-            ${TAB_ACTIONS[tab] ? html`
-              <button type="button" class="new-btn"
-                @click=${() => this[TAB_ACTIONS[tab].opener]()}
-              >${TAB_ACTIONS[tab].btnLabel}</button>
-            ` : nothing}
-            <button type="button"
-              class="chat-toggle-btn ${this._isChatOpen ? 'is-active' : ''}"
-              aria-label="${this._isChatOpen ? 'Close chat' : 'Open chat'}"
-              aria-pressed="${this._isChatOpen}"
-              @click=${() => this._toggleChat()}
-            >${this._isChatOpen ? 'Close Chat' : 'Chat'}</button>
-          </div>
-        </div>
-        ${showSearch ? html`
-          <div class="list-search">
-            <input
-              type="search"
-              placeholder="Search\u2026"
-              aria-label="Search list"
-              .value=${this._promptSearch}
-              @input=${(e) => { this._promptSearch = e.target.value; }}
-            >
-          </div>
-        ` : nothing}
-        <div class="catalog-scroll">
-          ${tab === 'skills' ? this._renderSkillsCatalog() : nothing}
-          ${tab === 'agents' ? this._renderAgentsCatalog() : nothing}
-          ${tab === 'prompts' ? this._renderPromptsCatalog() : nothing}
-          ${tab === 'mcps' ? this._renderMcpsCatalog() : nothing}
-          ${tab === 'memory' ? html`<div class="empty">Memory is shown in the panel \u2192</div>` : nothing}
-        </div>
-      </div>
-    `;
-  }
-
-  /** Derive the editor drawer title from the current tab and form state. */
-  _editorTitle(tab) {
-    if (tab === 'agents' && this._isAgentViewTools) return 'Associated Tools';
-    if (tab === 'agents') return this._isFormEdit ? 'Edit Agent' : 'New Agent';
-    if (tab === 'skills') return this._isFormEdit ? 'Edit Skill' : 'New Skill';
-    if (tab === 'prompts') return this._isFormPromptEdit ? 'Edit Prompt' : 'New Prompt';
-    if (tab === 'mcps') {
-      if (this._viewingMcpServerId && !this._editingMcpKey) return this._viewingMcpServerId;
-      if (this._editingMcpKey) return `Edit: ${this._editingMcpKey}`;
-      return 'Register MCP Server';
-    }
-    if (tab === 'memory') return 'Project Memory';
-    return '';
-  }
-
-  // ─── render: editor panel ─────────────────────────────────────────────────
-
-  _renderEditorPanel() {
-    const tab = this._catalogTab;
-    const isSkill = tab === 'skills';
-    const isPrompt = tab === 'prompts';
-    const isMcp = tab === 'mcps';
-    const isAgent = tab === 'agents';
-    const isMemory = tab === 'memory';
-
-    const title = this._editorTitle(tab);
-
-    return html`
-      <div class="col-editor" aria-hidden=${this._isEditorOpen ? 'false' : 'true'}
-        ?inert=${!this._isEditorOpen}>
-        <div class="col-editor-inner">
-          ${this._isEditorOpen ? html`
-            <div class="editor-header">
-              <h3 class="editor-title">${title}</h3>
-              <button type="button" class="btn-icon close-btn" aria-label="Close"
-                @click=${() => this._closeEditor()}
-              >\u2715</button>
-            </div>
-            ${this._isFormDirty ? html`
-              <div class="dirty-notice" role="status">Unsaved edits &middot; save to persist</div>
-            ` : nothing}
-            <div class="editor-body ${isMemory ? 'editor-body-memory' : ''}">
-              ${isSkill ? this._renderSkillForm() : nothing}
-              ${isAgent && this._isAgentViewTools ? this._renderAssociatedToolsSelector() : nothing}
-              ${isAgent && !this._isAgentViewTools ? this._renderAgentForm() : nothing}
-              ${isPrompt ? this._renderPromptForm() : nothing}
-              ${isMcp && (this._editingMcpKey || !this._viewingMcpServerId)
-                ? this._renderMcpForm() : nothing}
-              ${isMcp && this._viewingMcpServerId && !this._editingMcpKey
-                ? this._renderMcpServerInfo() : nothing}
-              ${isMcp && (this._viewingMcpServerId || this._editingMcpKey)
-                ? this._renderMcpToolsList() : nothing}
-              ${isMemory ? html`
-                <p class="form-hint">.da/agent/memory.md</p>
-                ${this._renderMemoryContent()}
-              ` : nothing}
-            </div>
-            ${(isSkill || (isAgent && !this._isAgentViewTools) || isPrompt
-              || (isMcp && (!this._viewingMcpServerId || this._editingMcpKey))) ? html`
-              <div class="editor-footer">
-                ${this._renderEditorFooter(isSkill, isPrompt, isMcp, isAgent)}
-              </div>
-            ` : nothing}
-          ` : nothing}
-        </div>
-      </div>
-    `;
-  }
-
-  // ─── render: chat drawer ──────────────────────────────────────────────────
-
-  _renderChatDrawer() {
-    return html`
-      <div class="chat-drawer" aria-hidden=${this._isChatOpen ? 'false' : 'true'}
-        ?inert=${!this._isChatOpen}>
-        ${this._isChatOpen ? html`
-          <div class="chat-drawer-header">
-            <span class="chat-drawer-title">Chat</span>
-            <button type="button" class="btn-icon close-btn" aria-label="Close chat"
-              @click=${() => this._toggleChat()}
-            >\u2715</button>
-          </div>
-          <div class="chat-drawer-body"
-            @nx-panel-close=${(e) => { e.stopPropagation(); this._toggleChat(); }}>
-            <nx-chat></nx-chat>
-          </div>
-        ` : nothing}
-      </div>
-    `;
-  }
-
-  // ─── render: skill form ───────────────────────────────────────────────────
-
-  _renderSkillForm() {
-    return html`
-      <form class="form" @submit=${(e) => e.preventDefault()}>
-        <input
-          type="text"
-          placeholder="skill-id"
-          aria-label="Skill ID"
-          .value=${this._formSkillId}
-          ?readonly=${this._isFormEdit}
-          @input=${(e) => { this._formSkillId = e.target.value; this._markDirty(); }}
-        >
-        <div class="textarea-wrap ${this._hasSuggestion ? 'is-suggestion' : ''}">
-          <textarea
-            placeholder="Write or revise skill markdown"
-            aria-label="Skill markdown"
-            .value=${this._formSkillBody}
-            @input=${(e) => { this._formSkillBody = e.target.value; this._markDirty(); }}
-          ></textarea>
-        </div>
-      </form>
-    `;
-  }
-
-  _renderAgentForm() {
-    return html`
-      <form class="form" @submit=${(e) => e.preventDefault()}>
-        <p class="form-hint">Creates <code>/.da/agents/&lt;id&gt;.json</code></p>
-        <input
-          type="text"
-          placeholder="agent-id"
-          aria-label="Agent ID"
-          .value=${this._newAgentId}
-          @input=${(e) => { this._newAgentId = e.target.value; this._markDirty(); }}
-        >
-        <input
-          type="text"
-          placeholder="Display name"
-          aria-label="Agent display name"
-          .value=${this._newAgentName}
-          @input=${(e) => { this._newAgentName = e.target.value; this._markDirty(); }}
-        >
-      </form>
-    `;
-  }
-
-  // ─── render: prompt form ──────────────────────────────────────────────────
-
-  _renderPromptForm() {
-    return html`
-      <form class="form" @submit=${(e) => e.preventDefault()}>
-        <input type="text" placeholder="Title" aria-label="Prompt title"
-          .value=${this._formPromptTitle}
-          @input=${(e) => { this._formPromptTitle = e.target.value; this._markDirty(); }}
-        >
-        <input type="text" placeholder="Category (e.g. Review, Workflow\u2026)" aria-label="Prompt category"
-          list="category-list"
-          .value=${this._formPromptCategory}
-          @input=${(e) => { this._formPromptCategory = e.target.value; this._markDirty(); }}
-        >
-        <input type="url" placeholder="Icon URL" aria-label="Prompt icon URL"
-          .value=${this._formPromptIcon}
-          @input=${(e) => { this._formPromptIcon = e.target.value; this._markDirty(); }}
-        >
-        <datalist id="category-list">
-          ${CATEGORY_OPTIONS.map((c) => html`<option value=${c}></option>`)}
-        </datalist>
-        <div class="textarea-wrap">
-          <textarea
-            placeholder="Write your prompt\u2026"
-            aria-label="Prompt body"
-            .value=${this._formPromptBody}
-            @input=${(e) => { this._formPromptBody = e.target.value; this._markDirty(); }}
-          ></textarea>
-        </div>
-      </form>
-    `;
-  }
-
-  // ─── render: associated tools selector ───────────────────────────────────
-
-  _renderAssociatedToolsSelector() {
-    const builtIn = BUILTIN_TOOL_IDS;
-    const mcpToolIds = [];
-    if (this._mcpTools?.servers) {
-      this._mcpTools.servers.forEach((server) => {
-        (server.tools || []).forEach((tool) => {
-          mcpToolIds.push(`mcp__${server.id}__${tool.name}`);
-        });
-      });
-    }
-
-    const toolFilter = (this._toolsSearch || '').trim().toLowerCase();
-    const filterById = (id) => id.toLowerCase().includes(toolFilter);
-    const daTools = toolFilter ? builtIn.filter(filterById) : builtIn;
-    const mcpTools = toolFilter ? mcpToolIds.filter(filterById) : mcpToolIds;
-    const selected = new Set(this._formPromptTools || []);
-    const collapsed = this._toolsGroupCollapsed || {};
-
-    const renderGroup = (ns, tools) => {
-      if (!tools.length && !toolFilter) return nothing;
-      const isOpen = !collapsed[ns];
-      return html`
-        <details class="tools-group" ?open=${isOpen}
-          @toggle=${(e) => {
-            this._toolsGroupCollapsed = { ...this._toolsGroupCollapsed, [ns]: !e.target.open };
-          }}
-        >
-          <summary class="tools-group-summary">
-            <span class="tools-group-label">${ns}</span>
-            <span class="tools-count">${tools.length}</span>
-          </summary>
-          <ul class="tools-group-list" aria-label="${ns} tools">
-            ${!tools.length ? html`<li class="tool-item-empty">No tools match filter</li>` : nothing}
-            ${tools.map((toolId) => {
-              const isActive = selected.has(toolId);
-              return html`
-                <li class="tool-item ${isActive ? 'is-active' : ''}">
-                  <span class="tool-dot ${isActive ? 'is-dot-active' : 'is-dot-inactive'}" aria-hidden="true"></span>
-                  <label class="tool-label-wrap" title=${toolId}>
-                    <input type="checkbox" class="tool-checkbox"
-                      .checked=${isActive}
-                      @change=${(e) => {
-                        const prevTools = this._formPromptTools ? [...this._formPromptTools] : [];
-                        const next = new Set(prevTools);
-                        if (e.target.checked) next.add(toolId);
-                        else next.delete(toolId);
-                        this._formPromptTools = [...next];
-                        const { serverId, toolName } = this._parseToolId(toolId);
-                        this._onToggleToolEnabled(serverId, toolName, e.target.checked, () => {
-                          this._formPromptTools = prevTools;
-                        });
-                      }}
-                    >
-                    <span class="tool-label">${toolId}</span>
-                  </label>
-                </li>
-              `;
-            })}
-          </ul>
-        </details>
-      `;
-    };
-
-    return html`
-      <div class="tools-selector">
-        <h4 class="tools-selector-heading">Associated Tools</h4>
-        <input
-          type="search"
-          class="tools-search-input"
-          placeholder="Filter tools\u2026"
-          aria-label="Filter tools"
-          .value=${this._toolsSearch}
-          @input=${(e) => { this._toolsSearch = e.target.value; }}
-        >
-        ${renderGroup('DA', daTools)}
-        ${mcpTools.length || toolFilter ? renderGroup('MCP', mcpTools) : nothing}
-      </div>
-    `;
-  }
-
-  // ─── render: MCP form ─────────────────────────────────────────────────────
-
-  _renderMcpForm() {
-    const hasSecret = Boolean(String(this._mcpAuthHeaderValue || '').trim());
-    return html`
-      <form class="form" @submit=${(e) => e.preventDefault()}>
-        <input type="text" placeholder="server-id (not API key)" aria-label="MCP server id"
-          .value=${this._mcpKey}
-          ?readonly=${Boolean(this._editingMcpKey)}
-          @input=${(e) => { this._mcpKey = e.target.value; this._markDirty(); }}
-        >
-        <p class="form-hint">Identifier only. Do not paste secrets or API keys here.</p>
-        <input type="text" placeholder="SSE endpoint URL" aria-label="MCP server URL"
-          .value=${this._mcpUrl}
-          @input=${(e) => { this._mcpUrl = e.target.value; this._markDirty(); }}
-        >
-        <textarea
-          class="textarea-sm"
-          placeholder="Description \u2014 what this server does (optional)"
-          aria-label="MCP server description"
-          .value=${this._mcpDescription}
-          @input=${(e) => { this._mcpDescription = e.target.value; this._markDirty(); }}
-        ></textarea>
-        <div class="mcp-auth-section ${hasSecret ? 'is-sensitive' : ''}">
-          <p class="form-hint">Authentication header (optional, for private MCP servers)</p>
-          <input
-            type="text"
-            placeholder="Header name (e.g. Authorization, x-api-key)"
-            aria-label="MCP auth header name"
-            .value=${this._mcpAuthHeaderName}
-            @input=${(e) => { this._mcpAuthHeaderName = e.target.value; this._markDirty(); }}
-          >
-          <input
-            type="password"
-            autocomplete="new-password"
-            placeholder="Header value"
-            aria-label="MCP auth header value"
-            .value=${this._mcpAuthHeaderValue}
-            @input=${(e) => { this._mcpAuthHeaderValue = e.target.value; this._markDirty(); }}
-          >
-          ${hasSecret ? html`
-            <p class="mcp-auth-warning" role="note">
-              ⚠ Saving this key makes it available to all authors with configuration permission.
-            </p>
-          ` : nothing}
-        </div>
-      </form>
-    `;
-  }
-
-  // ─── render: MCP server info (read-only, built-in) ────────────────────────
-
-  _renderMcpServerInfo() {
-    const serverId = this._viewingMcpServerId;
-    const builtin = BUILTIN_MCP_SERVERS.find((s) => s.id === serverId);
-    if (!builtin) return nothing;
-    return html`
-      <div class="mcp-server-info">
-        <p class="mcp-server-desc">${builtin.description}</p>
-        <span class="badge">built-in</span>
-      </div>
-    `;
-  }
-
-  // ─── render: MCP tools list ──────────────────────────────────────────────
-
-  _mcpServerToolData(serverId) {
-    const builtinList = BUILTIN_TOOL_DETAILS[serverId];
-    if (builtinList) return { tools: builtinList, error: null, source: 'builtin' };
-
-    if (!this._mcpTools) return { tools: [], error: null, source: 'pending' };
-
-    const server = (this._mcpTools.servers || []).find((s) => s.id === serverId);
-    if (!server) {
-      const isConfigured = Boolean(this._configuredMcpServers?.[serverId]);
-      if (!isConfigured) return { tools: [], error: 'Server is disabled or has no URL', source: 'unconfigured' };
-      return { tools: [], error: null, source: 'pending' };
-    }
-
-    if (server.error) return { tools: [], error: server.error, source: 'error' };
-    const tools = (server.tools || []).map((t) => ({
-      name: t.name,
-      description: t.description || '',
-    }));
-    return { tools, error: null, source: 'live' };
-  }
-
-  _renderMcpToolsList() {
-    const serverId = this._viewingMcpServerId || this._editingMcpKey;
-    if (!serverId) return nothing;
-
-    const { tools, error, source } = this._mcpServerToolData(serverId);
-
-    const overrides = this._toolOverrides || {};
-    const filterQ = (this._toolsSearch || '').trim().toLowerCase();
-    const filtered = filterQ
-      ? tools.filter((t) => t.name.toLowerCase().includes(filterQ)
-        || t.description.toLowerCase().includes(filterQ))
-      : tools;
-
-    const emptyMsg = () => {
-      if (source === 'pending') return 'Connecting to agent to discover tools\u2026';
-      if (source === 'unconfigured') return 'Enable this server to discover its tools';
-      if (source === 'error') {
-        const urlMatch = error?.match(/https?:\/\/\S+/);
-        const hint = urlMatch?.[0];
-        const base = error?.split('\n')[0] ?? error;
-        return html`
-          Could not list tools: ${base}
-          ${hint ? html`
-            <br>
-            <span class="mcp-error-hint">Did you mean:
-              <a class="mcp-error-url" href="#"
-                @click=${(e) => {
-                  e.preventDefault();
-                  this._mcpUrl = hint;
-                  this._setStatus(`URL updated to ${hint} — save to apply`, STATUS_TYPE.WARN);
-                }}
-              >${hint}</a>?
-            </span>
-          ` : nothing}
-        `;
-      }
-      return 'Server reported 0 tools';
-    };
-
-    return html`
-      <div class="mcp-tools-section">
-        <h4 class="tools-selector-heading">Tools (${tools.length})</h4>
-        ${tools.length > 6 ? html`
-          <input type="search" class="tools-search-input"
-            placeholder="Filter tools\u2026" aria-label="Filter tools"
-            .value=${this._toolsSearch}
-            @input=${(e) => { this._toolsSearch = e.target.value; }}
-          >
-        ` : nothing}
-        ${!tools.length
-          ? html`<div class="empty ${source === 'error' ? 'empty-err' : ''}">${emptyMsg()}</div>`
-          : html`
-            <ul class="tools-group-list" aria-label="Tools for ${serverId}">
-              ${filtered.map((t) => {
-                const key = `${serverId}/${t.name}`;
-                const isEnabled = overrides[key] !== false;
-                return html`
-                  <li class="tool-item ${isEnabled ? 'is-active' : ''}">
-                    <label class="tool-label-wrap" title=${t.name}>
-                      <input type="checkbox" class="tool-checkbox"
-                        .checked=${isEnabled}
-                        @change=${(e) => this._onToggleToolEnabled(serverId, t.name, e.target.checked)}
-                      >
-                      <div class="tool-text">
-                        <span class="tool-label">${t.name}</span>
-                        ${t.description ? html`
-                          <span class="tool-desc">${t.description}</span>
-                        ` : nothing}
-                      </div>
-                    </label>
-                  </li>
-                `;
-              })}
-              ${filtered.length === 0 && tools.length
-                ? html`<li class="tool-item-empty">No tools match filter</li>` : nothing}
-            </ul>
-          `}
-      </div>
-    `;
-  }
-
-  // ─── render: editor footer (sticky actions) ───────────────────────────────
-
-  _renderEditorFooter(isSkill, isPrompt, isMcp, isAgent) {
-    const statusTpl = this._statusMsg ? html`
-      <output class="msg ${this._msgClass()}">
-        ${this._statusMsg}
-      </output>
-    ` : nothing;
-
-    if (isSkill) {
-      return html`
-        <div class="editor-actions" role="toolbar" aria-label="Skill actions">
-          ${this._isFormEdit || this._hasSuggestion ? html`
-            <button type="button" data-variant="secondary"
-              ?disabled=${this._isSaveBusy}
-              @click=${() => { this._dismissForm(); }}
-            >Dismiss</button>
-          ` : nothing}
-          <button type="button" data-variant="secondary"
-            ?disabled=${this._isSaveBusy}
-            @click=${() => this._onSaveSkill(STATUS.DRAFT)}
-          >Save Draft</button>
-          <button type="button" data-variant="accent"
-            ?disabled=${this._isSaveBusy}
-            @click=${() => this._onSaveSkill(STATUS.APPROVED)}
-          >Save</button>
-          ${this._isFormEdit ? html`
-            <button type="button" data-variant="negative"
-              ?disabled=${this._isSaveBusy}
-              @click=${this._onDeleteSkill}
-            >Delete</button>
-          ` : nothing}
-        </div>
-        ${statusTpl}
-      `;
-    }
-
-    if (isAgent) {
-      return html`
-        <div class="editor-actions" role="toolbar" aria-label="Agent actions">
-          <button type="button" data-variant="accent"
-            ?disabled=${this._isSaveBusy || !this._newAgentId.trim()}
-            @click=${this._onSaveAgent}
-          >Save Agent File</button>
-        </div>
-        ${statusTpl}
-      `;
-    }
-
-    if (isPrompt) {
-      return html`
-        <div class="editor-actions" role="toolbar" aria-label="Prompt actions">
-          <button type="button" data-variant="secondary"
-            ?disabled=${this._isSaveBusy}
-            @click=${() => this._onSavePrompt(STATUS.DRAFT)}
-          >Save Draft</button>
-          <button type="button" data-variant="accent"
-            ?disabled=${this._isSaveBusy}
-            @click=${() => this._onSavePrompt(STATUS.APPROVED)}
-          >Save</button>
-          <button type="button" data-variant="secondary"
-            ?disabled=${this._isSaveBusy || !this._formPromptBody.trim()}
-            @click=${() => {
-              this._dispatchPromptToChat(DA_SKILLS_EDITOR_PROMPT_ADD_TO_CHAT, this._formPromptBody);
-              this._dispatchPromptToChat(DA_SKILLS_LAB_PROMPT_ADD_TO_CHAT, this._formPromptBody);
-              this._setStatus('Added to chat input');
-            }}
-          >Add to Chat</button>
-          <button type="button" data-variant="secondary"
-            ?disabled=${this._isSaveBusy || !this._formPromptBody.trim()}
-            @click=${() => this._onRunPrompt()}
-          >Run / Test</button>
-          ${this._isFormPromptEdit ? html`
-            <button type="button" data-variant="negative"
-              ?disabled=${this._isSaveBusy}
-              @click=${this._onDeletePrompt}
-            >Delete</button>
-          ` : nothing}
-        </div>
-        ${statusTpl}
-      `;
-    }
-
-    if (isMcp) {
-      return html`
-        <div class="editor-actions" role="toolbar" aria-label="MCP actions">
-          <button type="button" data-variant="accent"
-            ?disabled=${this._isSaveBusy || !this._mcpKey.trim() || !this._mcpUrl.trim()}
-            @click=${this._onRegisterMcp}
-          >${this._editingMcpKey ? 'Update' : 'Register'}</button>        </div>
-        ${statusTpl}
-      `;
-    }
-
-    return nothing;
-  }
-
-  // ─── render: skills catalog ───────────────────────────────────────────────
-
-  _renderSkillsCatalog() {
-    const ids = Object.keys(this._skills);
-    const searchQuery = this._promptSearch.trim().toLowerCase();
-
-    let filtered = this._catalogFilter === 'all' ? ids
-      : ids.filter((id) => this._skillStatuses[id] === this._catalogFilter);
-
-    if (searchQuery) {
-      filtered = filtered.filter((id) => {
-        const title = this._extractTitle(this._skills[id]).toLowerCase();
-        return id.toLowerCase().includes(searchQuery) || title.includes(searchQuery);
-      });
-    }
-
-    return html`
-      <div class="catalog-toolbar" role="toolbar" aria-label="Filter skills">
-        ${[STATUS.APPROVED, STATUS.DRAFT].map((status) => html`
-          <button type="button"
-            class="filter-chip ${this._catalogFilter === status ? 'is-active' : ''}"
-            aria-pressed=${this._catalogFilter === status ? 'true' : 'false'}
-            @click=${() => { this._catalogFilter = status; }}
-          >${status.charAt(0).toUpperCase() + status.slice(1)}</button>
-        `)}
-        <button type="button"
-          class="filter-chip ${this._catalogFilter === 'all' ? 'is-active' : ''}"
-          aria-pressed=${this._catalogFilter === 'all' ? 'true' : 'false'}
-          @click=${() => { this._catalogFilter = 'all'; }}
-        >All</button>
-      </div>
-      ${!filtered.length
-        ? html`<div class="empty">No skills found</div>`
-        : filtered.map((id) => this._renderSkillCard(id))}
-    `;
-  }
-
-  _renderSkillCard(id) {
-    const title = this._extractTitle(this._skills[id]);
-    const status = this._skillStatuses[id] || STATUS.APPROVED;
-    const isEditing = this._isFormEdit && this._formSkillId === id;
-    const isDraft = status === STATUS.DRAFT;
-    return html`
-      <article
-        role="button"
-        tabindex="0"
-        aria-label="Edit skill ${id}"
-        data-testid="skill-card"
-        data-skill-id=${id}
-        @click=${(e) => this._onCardClick(e, () => this._onEditSkill(id))}
-        @keydown=${(e) => this._onCardKeydown(e, () => this._onEditSkill(id))}
-      >
-        <nx-card
-          interactive
-          heading=${id}
-          subheading=${title || nothing}
-          ?selected=${isEditing}
-        >
-          <span slot="pill"
-            class="status-dot ${isDraft ? 'status-dot-draft' : 'status-dot-approved'}"
-            aria-label=${isDraft ? 'Draft' : 'Approved'}
-          ></span>
-          <button slot="actions" type="button" class="btn-icon more-btn"
-            aria-label="More actions for ${id}"
-            @click=${(e) => { e.stopPropagation(); this._openSkillMenu(e, id); }}
-          >\u22ee</button>
-        </nx-card>
-        <nx-popover placement="auto">
-          <div class="card-menu" role="menu">
-            <button role="menuitem" type="button"
-              @click=${() => { this._closeSkillMenu(id); this._onEditSkill(id); }}
-            >Edit</button>
-            <button role="menuitem" type="button" class="card-menu-delete"
-              @click=${() => { this._closeSkillMenu(id); this._onDeleteSkillById(id); }}
-            >Delete</button>
-          </div>
-        </nx-popover>
-      </article>
-    `;
-  }
-
-  // ─── render: agents catalog ───────────────────────────────────────────────
-
-  _renderAgentCard(agent, isBuiltin = false) {
-    const title = agent.label || agent.name || agent.preset?.name || agent.id;
-    const description = agent.description || agent.preset?.description || '';
-    const tools = this._agentToolIds(agent, isBuiltin);
-    return html`
-      <article class="agent-card" role="button" tabindex="0"
-        aria-label="Open agent ${title}"
-        data-testid=${isBuiltin ? 'agent-builtin-card' : 'agent-card'}
-        @click=${(e) => this._onCardClick(e, () => this._onSelectAgent(agent))}
-        @keydown=${(e) => this._onCardKeydown(e, () => this._onSelectAgent(agent))}
-      >
-        <header class="agent-card-header">
-          <span class="status-dot status-dot-approved" aria-label="Active"></span>
-          <span class="agent-card-title">${title}</span>
-          <span class="badge">${isBuiltin ? 'built-in' : 'custom'}</span>
-        </header>
-        ${description ? html`<p class="agent-card-desc">${description}</p>` : nothing}
-        ${tools.length ? html`
-          <footer class="agent-card-footer">
-            <ul class="agent-tools-list" aria-label="Tools used by ${title}">
-              ${tools.slice(0, 12).map((tool) => html`<li class="agent-tool-chip">${tool}</li>`)}
-            </ul>
-          </footer>
-        ` : nothing}
-      </article>
-    `;
-  }
-
-  _renderAgentsCatalog() {
-    return html`
-      <h3 class="section-h">Built-in (${BUILTIN_AGENTS.length})</h3>
-      ${BUILTIN_AGENTS.map((agent) => this._renderAgentCard(agent, true))}
-      ${this._agents.length ? html`
-        <h3 class="section-h">Custom (${this._agents.length})</h3>
-        ${this._agents.map((agent) => this._renderAgentCard(agent, false))}
-      ` : nothing}
-      ${this._agentRows.length ? html`
-        <h3 class="section-h">Config Agents (${this._agentRows.length})</h3>
-        ${this._agentRows.map((row) => html`
-          <article class="agent-card" role="listitem" data-testid="agent-config-card">
-            <header class="agent-card-header">
-              <span class="status-dot status-dot-approved" aria-label="Configured"></span>
-              <span class="agent-card-title">${row.key}</span>
-              <span class="badge">config</span>
-            </header>
-            <p class="agent-card-desc">${row.url}</p>
-          </article>
-        `)}
-      ` : nothing}
-    `;
-  }
-
-  // ─── render: prompts catalog ──────────────────────────────────────────────
-
-  _renderPromptsCatalog() {
-    const searchQuery = this._promptSearch.trim().toLowerCase();
-    const prompts = searchQuery
-      ? this._prompts.filter((r) => (r.title || '').toLowerCase().includes(searchQuery)
-        || (r.category || '').toLowerCase().includes(searchQuery))
-      : this._prompts;
-
-    if (!prompts.length) {
-      return html`<div class="empty">No prompts found</div>`;
-    }
-
-    return html`
-      <div role="list" aria-label="Prompts">
-        ${prompts.map((row) => {
-          const title = row.title || '';
-          const isSelected = this._isEditorOpen && this._isFormPromptEdit
-            && this._formPromptTitle === title;
-          const cat = (row.category || '').toLowerCase().trim();
-          const catClass = KNOWN_CATEGORY_CLASSES.has(cat) ? cat : 'default';
-          return html`
-            <article role="listitem" data-testid="prompt-card" data-prompt-title=${title}>
-              <div class="prompt-row ${isSelected ? 'is-selected' : ''}" role="button"
-                tabindex="0"
-                aria-label="Edit prompt ${title || '(untitled)'}"
-                @click=${(e) => this._onCardClick(e, () => this._openEditor(row))}
-                @keydown=${(e) => this._onCardKeydown(e, () => this._openEditor(row))}
-              >
-                <div class="prompt-row-body">
-                  <span class="prompt-row-title">${title || '(untitled)'}</span>
-                  ${row.category ? html`
-                    <span class="category-badge cat-${catClass}">${row.category}</span>
-                  ` : nothing}
-                </div>
-                <div class="prompt-row-actions">
-                  <button type="button" class="btn-icon row-action-btn" title="Edit"
-                    aria-label="Edit ${title}"
-                    @click=${(e) => { e.stopPropagation(); this._openEditor(row); }}
-                  >\u270e</button>
-                  <button type="button" class="btn-icon row-action-btn" title="Duplicate"
-                    aria-label="Duplicate ${title}"
-                    @click=${(e) => { e.stopPropagation(); this._duplicatePrompt(row); }}
-                  >\u29c9</button>
-                  <button type="button" class="btn-icon row-action-btn" title="Add to chat"
-                    aria-label="Add to chat: ${title}"
-                    @click=${(e) => {
-                      e.stopPropagation();
-                      this._dispatchPromptToChat(DA_SKILLS_EDITOR_PROMPT_ADD_TO_CHAT, row.prompt);
-                      this._dispatchPromptToChat(DA_SKILLS_LAB_PROMPT_ADD_TO_CHAT, row.prompt);
-                    }}
-                  >+</button>
-                  <button type="button" class="btn-icon row-action-btn" title="Send to chat"
-                    aria-label="Send to chat: ${title}"
-                    @click=${(e) => {
-                      e.stopPropagation();
-                      this._dispatchPromptToChat(DA_SKILLS_EDITOR_PROMPT_SEND, row.prompt);
-                      this._dispatchPromptToChat(DA_SKILLS_LAB_PROMPT_SEND, row.prompt);
-                    }}
-                  >\u25b6</button>
-                  <button type="button" class="btn-icon row-action-btn row-action-btn-delete" title="Delete"
-                    aria-label="Delete ${title}"
-                    @click=${(e) => { e.stopPropagation(); this._deletePromptDirect(row); }}
-                  >\u{1F5D1}</button>
-                </div>
-              </div>
-            </article>
-          `;
-        })}
-      </div>
-    `;
-  }
-
-  // ─── render: MCPs catalog ─────────────────────────────────────────────────
-
-  _renderMcpsCatalog() {
-    const searchQuery = this._promptSearch.trim().toLowerCase();
-    const filterPasses = (status) => this._catalogFilter === 'all' || status === this._catalogFilter;
-    let filteredCustom = this._mcpRows.filter((row) => filterPasses(skillRowStatus(row)));
-    if (searchQuery) {
-      filteredCustom = filteredCustom.filter((row) => {
-        const key = (row.key || '').toLowerCase();
-        const url = (row.url || row.value || '').toLowerCase();
-        return key.includes(searchQuery) || url.includes(searchQuery);
-      });
-    }
-    const showBuiltins = filterPasses(STATUS.APPROVED);
-
-    return html`
-      ${showBuiltins ? html`
-        <h3 class="section-h">Built-in (${BUILTIN_MCP_SERVERS.length})</h3>
-        ${BUILTIN_MCP_SERVERS.map((s) => {
-          const isViewing = this._viewingMcpServerId === s.id && !this._editingMcpKey;
-          return html`
-            <article
-              role="button"
-              tabindex="0"
-              aria-label="View tools for ${s.id}"
-              data-testid="mcp-builtin-card"
-              @click=${(e) => this._onMcpCardClick(e, () => this._onViewMcpTools(s.id))}
-              @keydown=${(e) => this._onMcpCardKeydown(e, () => this._onViewMcpTools(s.id))}
-            >
-              <nx-card heading=${s.id} subheading=${s.description}
-                interactive
-                ?selected=${isViewing}>
-                <span slot="pill" class="status-dot status-dot-approved"
-                  aria-label="Enabled"></span>
-                <span slot="actions" class="badge">built-in</span>
-              </nx-card>
-            </article>
-          `;
-        })}
-      ` : nothing}
-      <h3 class="section-h">Custom (${filteredCustom.length})</h3>
-      ${!filteredCustom.length
-        ? html`<div class="empty">No custom MCP servers registered</div>`
-        : filteredCustom.map((row) => {
-          const isApproved = skillRowStatus(row) === STATUS.APPROVED;
-          const isEnabled = isApproved && skillRowEnabled(row);
-          const key = row.key || '';
-          const token = `mcp:${key}`;
-          const isBusy = this._mcpEnableBusy[token];
-          const isSelected = this._isEditorOpen
-            && (this._editingMcpKey === key || this._viewingMcpServerId === key);
-          return html`
-            <article
-              role="button"
-              tabindex="0"
-              aria-label="Edit MCP server ${key || '(unnamed)'}"
-              data-testid="mcp-card"
-              data-mcp-key=${key}
-              @click=${(e) => this._onMcpCardClick(e, () => this._onEditMcp(row))}
-              @keydown=${(e) => this._onMcpCardKeydown(e, () => this._onEditMcp(row))}
-            >
-              <nx-card heading=${key || '(unnamed)'}
-                interactive
-                subheading=${row.description || row.url || row.value || ''}
-                ?selected=${isSelected}>
-                <span slot="pill"
-                  class="status-dot ${isEnabled ? 'status-dot-approved' : 'status-dot-draft'}"
-                  aria-label=${isEnabled ? 'Enabled' : 'Disabled'}
-                ></span>
-                <button slot="actions" type="button" class="btn-icon more-btn"
-                  aria-label="More actions for ${key}"
-                  @click=${(e) => { e.stopPropagation(); this._openMcpMenu(e, key); }}
-                >\u22ee</button>
-              </nx-card>
-              <nx-popover placement="auto">
-                <div class="card-menu" role="menu"
-                  @click=${(e) => e.stopPropagation()}>
-                  ${isApproved ? html`
-                    <button role="menuitem" type="button"
-                      ?disabled=${isBusy}
-                      @click=${() => { this._closeMcpMenu(key); this._onToggleMcpEnabled(row); }}
-                    >${isEnabled ? 'Disable' : 'Enable'}</button>
-                  ` : nothing}
-                  <button role="menuitem" type="button"
-                    @click=${() => { this._closeMcpMenu(key); this._onEditMcp(row); }}
-                  >Edit</button>
-                  <button role="menuitem" type="button" class="card-menu-delete"
-                    @click=${() => { this._closeMcpMenu(key); this._onDeleteMcpDirect(row); }}
-                  >Delete</button>
-                </div>
-              </nx-popover>
-            </article>
-          `;
-        })}
-    `;
-  }
-
-  // ─── render: memory catalog ───────────────────────────────────────────────
-
-  _renderMemoryCatalog() {
-    return html`
-      <h3 class="section-h">Project Memory</h3>
-      <p class="form-hint">.da/agent/memory.md</p>
-      ${this._renderMemoryContent()}
-    `;
-  }
-
-  _renderMemoryContent() {
-    if (this._memory === null) {
-      return html`<div class="empty" aria-live="polite">Loading\u2026</div>`;
-    }
-    if (this._memory === '') {
-      return html`<div class="empty">No project memory yet. The DA agent writes here as it learns about your site.</div>`;
-    }
-    return html`<pre class="memory-content">${this._memory}</pre>`;
-  }
-
-  // ─── utility ──────────────────────────────────────────────────────────────
-
-  _extractTitle(md) {
-    if (!md) return '';
-    const match = md.match(/^#\s+(.+)/m);
-    return match ? match[1].trim() : '';
   }
 }
